# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/authCode": {
            "post": {
                "description": "发送手机验证码",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "手机验证码"
                ],
                "summary": "发送手机验证码",
                "parameters": [
                    {
                        "type": "string",
                        "default": "13881887710",
                        "description": "手机号码",
                        "name": "mobile",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1",
                        "description": "发送类型:1登陆,2注册,3忘记密码",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JsonResult"
                        }
                    }
                }
            }
        },
        "/forgetPwd": {
            "post": {
                "description": "忘记密码",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户登陆"
                ],
                "summary": "忘记密码",
                "parameters": [
                    {
                        "type": "string",
                        "default": "13881887710",
                        "description": "手机号码",
                        "name": "mobile",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "123456",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "432343",
                        "description": "短信验证码",
                        "name": "sms_code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1614479454",
                        "description": "时间戳",
                        "name": "timestamp",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "5fae1a0e4156e1ad4e742f4b9b7c1416",
                        "description": "签名",
                        "name": "sign",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JsonResult"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "用户登陆",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户登陆"
                ],
                "summary": "用户登陆",
                "parameters": [
                    {
                        "type": "string",
                        "default": "13881887710",
                        "description": "手机号码",
                        "name": "mobile",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "123456",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "432343",
                        "description": "短信验证码",
                        "name": "sms_code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1614479454",
                        "description": "时间戳",
                        "name": "timestamp",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "5fae1a0e4156e1ad4e742f4b9b7c1416",
                        "description": "签名",
                        "name": "sign",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JsonResult"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户登陆"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "type": "string",
                        "default": "13881887710",
                        "description": "手机号码",
                        "name": "mobile",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "123456",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "432343",
                        "description": "短信验证码",
                        "name": "sms_code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "张三",
                        "description": "用户名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1",
                        "description": "性别:1男,2女",
                        "name": "sex",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1614479454",
                        "description": "时间戳",
                        "name": "timestamp",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "5fae1a0e4156e1ad4e742f4b9b7c1416",
                        "description": "签名",
                        "name": "sign",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JsonResult"
                        }
                    }
                }
            }
        },
...
        "/user/addrList": {
            "post": {
                "description": "获取收货地址列表",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取收货地址列表",
                "parameters": [
                    {
                        "type": "string",
                        "default": "eyJhbGciOiJIUzM4NCIsInR5cCI6IkpXVCJ9.eyJ1aWQiOjEsImV4cCI6MTYxNzA5MzgxNCwiaXNzIjoid2ViX2dpbl90ZW1wbGF0ZSIsIlJlZnJlc2hUaW1lIjoxNjE1Nzk3ODE0fQ.LmTyXCemtbMu7rGcxQtefEdU04RdVbp9hk2_6ulW8retAULYhVa5zvRBl8IXoFEm",
                        "description": "用户token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JsonResult"
                        }
                    }
                }
            }
        },
...
        "/user/getAlipay": {
            "post": {
                "description": "获取alipay",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取alipay",
                "parameters": [
                    {
                        "type": "string",
                        "default": "eyJhbGciOiJIUzM4NCIsInR5cCI6IkpXVCJ9.eyJ1aWQiOjEsImV4cCI6MTYxNzA5MzgxNCwiaXNzIjoid2ViX2dpbl90ZW1wbGF0ZSIsIlJlZnJlc2hUaW1lIjoxNjE1Nzk3ODE0fQ.LmTyXCemtbMu7rGcxQtefEdU04RdVbp9hk2_6ulW8retAULYhVa5zvRBl8IXoFEm",
                        "description": "用户token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.JsonResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.JsonResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}